+++ express.js
@@ -4,7 +4,7 @@ var webpack = require('webpack')
 
 var createServeReactAppConfig = require('./lib/createServeReactAppConfig')
 var createServeReactAppBuildConfig = require('./lib/createServeReactAppBuildConfig')
-var createServeReactWebpackConfig = require('./lib/createServeReactWebpackConfig')
+var createServerWebpackConfig = require('./lib/createServerWebpackConfig')
 var debug = require('./lib/debug')
 
 /**
@@ -12,13 +12,20 @@ var debug = require('./lib/debug')
  * having run `nwb serve`, but from your own server.
  */
 module.exports = function(express, options) {
-  assert(express, 'The express module must be passed as the first argument to nwb middleware')
+  assert(
+    express && typeof express.Router === 'function',
+    'The express module must be passed as the first argument to nwb middleware'
+  )
 
   if (!options) options = {}
-  if (!('info' in options)) options.info = false
 
-  var webpackConfig = createServeReactWebpackConfig(
-    options,
+  // Use options to create an object equivalent to CLI args parsed by minimist
+  var args = {}
+  args.info = !!options.info
+  args['auto-install'] = !!options.autoInstall
+
+  var webpackConfig = createServerWebpackConfig(
+    args,
     createServeReactAppBuildConfig(
       createServeReactAppConfig()
     )
@@ -31,7 +38,7 @@ module.exports = function(express, options) {
   var router = express.Router()
 
   router.use(require('webpack-dev-middleware')(compiler, {
-    noInfo: !options.info,
+    noInfo: !args.info,
     publicPath: webpackConfig.output.publicPath,
     stats: {
       colors: true